+++ hail/python/hail/expr/table_type.py
@@ -1,3 +1,4 @@
+import pprint
 from hail.typecheck import typecheck_method, sequenceof
 from hail.utils.java import escape_parsable
 from hail.expr.types import dtype, tstruct
@@ -44,29 +45,29 @@ class ttable(object):
         return f'table {{global: {self.global_type}, row: {self.row_type}, row_key: [{self._key_str()}]}}'
 
     def pretty(self, indent=0, increment=4):
-        l = []
-        l.append(' ' * indent)
-        l.append('table {\n')
+        b = []
+        b.append(' ' * indent)
+        b.append('table {\n')
         indent += increment
-        
-        l.append(' ' * indent)
-        l.append('global: ')
-        self.global_type._pretty(l, indent, increment)
-        l.append(',\n')
-        
-        l.append(' ' * indent)
-        l.append('row: ')
-        self.row_type._pretty(l, indent, increment)
-        l.append(',\n')
-        
-        l.append(' ' * indent)
-        l.append(f'row_key: [{self._key_str()}]\n')
+
+        b.append(' ' * indent)
+        b.append('global: ')
+        self.global_type._pretty(b, indent, increment)
+        b.append(',\n')
+
+        b.append(' ' * indent)
+        b.append('row: ')
+        self.row_type._pretty(b, indent, increment)
+        b.append(',\n')
+
+        b.append(' ' * indent)
+        b.append(f'row_key: [{self._key_str()}]\n')
 
         indent -= increment
-        l.append(' ' * indent)
-        l.append('}')
-        
-        return ''.join(l)
+        b.append(' ' * indent)
+        b.append('}')
+
+        return ''.join(b)
 
     @property
     def key_type(self):
@@ -93,8 +94,6 @@ class ttable(object):
         else:
             return {'global': default_value}
 
-
-import pprint
 
 _old_printer = pprint.PrettyPrinter