+++ hail/python/hailtop/batch_client/aioclient.py
@@ -67,6 +67,10 @@ class Job:
 
     @staticmethod
     def _get_container_status_exit_code(container_status):
+        error = container_status.get('error')
+        if error is not None:
+            return None
+
         docker_container_status = container_status.get('container_status')
         if not docker_container_status:
             return None
@@ -209,6 +213,9 @@ class Job:
     async def log(self):
         return await self._job.log()
 
+    async def attempts(self):
+        return await self._job.attempts()
+
 
 class UnsubmittedJob:
     def _submit(self, batch):
@@ -248,6 +255,9 @@ class UnsubmittedJob:
 
     async def log(self):
         raise ValueError("cannot get the log of an unsubmitted job")
+
+    async def attempts(self):
+        raise ValueError("cannot get the attempts of an unsubmitted job")
 
 
 class SubmittedJob:
@@ -290,6 +300,10 @@ class SubmittedJob:
 
     async def log(self):
         resp = await self._batch._client._get(f'/api/v1alpha/batches/{self.batch_id}/jobs/{self.job_id}/log')
+        return await resp.json()
+
+    async def attempts(self):
+        resp = await self._batch._client._get(f'/api/v1alpha/batches/{self.batch_id}/jobs/{self.job_id}/attempts')
         return await resp.json()